+++ MyCollection/Pages/Settings/BoneGrades/Delete.cshtml.cs
@@ -39,7 +39,7 @@ namespace MyCollection.Pages.Grades
             else 
             {
                 var user = await _userManager.GetUserAsync(User);
-                if (user == null || grade.User != user)
+                if (user == null || grade.User?.Id != user.Id)
                 {
                     return RedirectToPage("/AccessDenied");
                 }
@@ -58,6 +58,11 @@ namespace MyCollection.Pages.Grades
 
             if (grade != null)
             {
+                var user = await _userManager.GetUserAsync(User);
+                if (user == null || grade.User?.Id != user.Id)
+                {
+                    return RedirectToPage("/AccessDenied");
+                }
                 Grade = grade;
                 _context.BoneGrades.Remove(Grade);
                 await _context.SaveChangesAsync();
